+++ suzie_doctor/rootfs/app/suzie_doctor/__init__.py
@@ -1,5 +1,6 @@
 """Suzie Doctor App package."""
 
-APP_VERSION = "0.2.35-dev"
+APP_VERSION = "0.2.36-dev"
 PROTOCOL_PACK_VERSION = "0.1.4-dev"
 BRIDGE_VERSION = "0.2.4-dev"
+